fix the misplacement of code caused by go gen
