+++ pkg/sloop/store/typed/watchactivitytable_test.go
@@ -155,6 +155,35 @@ func Test_WatchActivity_getLastMatchingKeyInPartition_NotFound(t *testing.T) {
 	assert.Nil(t, err)
 }
 
+func Test_WatchActivity_GetPreviousKey_Success(t *testing.T) {
+	db, wt := helper_update_WatchActivityTable(t, (&WatchActivityKey{}).SetTestKeys(), (&WatchActivityKey{}).GetTestValue())
+	var partRes *WatchActivityKey
+	var err1 error
+	curKey := NewWatchActivityKey(someMaxPartition, someKind, someNamespace, someName, someUid+"c")
+	keyComarator := NewWatchActivityKeyComparator(someKind, someNamespace, someName, someUid)
+	err := db.View(func(txn badgerwrap.Txn) error {
+		partRes, err1 = wt.GetPreviousKey(txn, curKey, keyComarator)
+		return err1
+	})
+	assert.Nil(t, err)
+	expectedKey := NewWatchActivityKey(someMaxPartition, someKind, someNamespace, someName, someUid)
+	assert.Equal(t, expectedKey, partRes)
+}
+
+func Test_WatchActivity_GetPreviousKey_Fail(t *testing.T) {
+	db, wt := helper_update_WatchActivityTable(t, (&WatchActivityKey{}).SetTestKeys(), (&WatchActivityKey{}).GetTestValue())
+	var partRes *WatchActivityKey
+	var err1 error
+	curKey := NewWatchActivityKey(someMaxPartition, someKind, someNamespace, someName, someUid)
+	keyComarator := NewWatchActivityKeyComparator(someKind+"a", someNamespace, someName, someUid)
+	err := db.View(func(txn badgerwrap.Txn) error {
+		partRes, err1 = wt.GetPreviousKey(txn, curKey, keyComarator)
+		return err1
+	})
+	assert.NotNil(t, err)
+	assert.Equal(t, &WatchActivityKey{}, partRes)
+}
+
 func (*WatchActivityKey) GetTestKey() string {
 	k := NewWatchActivityKey(someMinPartition, someKind, someNamespace, someName, someUid)
 	return k.String()
