+++ pkg/sloop/store/typed/watchtablegen.go
@@ -157,7 +157,6 @@ func (t *KubeWatchResultTable) GetPreviousKey(txn badgerwrap.Txn, key *WatchTabl
 		return &WatchTableKey{}, errors.Wrapf(err, "failed to get partition list from table:%v", t.tableName)
 	}
 	currentPartition := key.PartitionId
-
 	for i := len(partitionList) - 1; i >= 0; i-- {
 		prePart := partitionList[i]
 		if prePart > currentPartition {
