+++ pkg/sloop/store/typed/watchtablegen_test.go
@@ -112,32 +112,3 @@ func Test_KubeWatchResultTable_GetUniquePartitionList_EmptyPartition(t *testing.
 	assert.Nil(t, err)
 	assert.Len(t, partList, 0)
 }
-
-func Test_GetPreviousKey_Success(t *testing.T) {
-	db, wt := helper_update_KubeWatchResultTable(t, (&WatchTableKey{}).SetTestKeys(), (&WatchTableKey{}).SetTestValue())
-	var partRes *WatchTableKey
-	var err1 error
-	curKey := NewWatchTableKey(someMaxPartition, someKind, someNamespace, someName, someTs)
-	keyComparator := NewWatchTableKeyComparator(someKind, someNamespace, someName, zeroData)
-	err := db.View(func(txn badgerwrap.Txn) error {
-		partRes, err1 = wt.GetPreviousKey(txn, curKey, keyComparator)
-		return err1
-	})
-	assert.Nil(t, err)
-	expectedKey := NewWatchTableKey(someMaxPartition, someKind, someNamespace, someName, someTs.Add(time.Hour*-5))
-	assert.Equal(t, expectedKey, partRes)
-}
-
-func Test_GetPreviousKey_Fail(t *testing.T) {
-	db, wt := helper_update_KubeWatchResultTable(t, (&WatchTableKey{}).SetTestKeys(), (&WatchTableKey{}).SetTestValue())
-	var partRes *WatchTableKey
-	var err1 error
-	curKey := NewWatchTableKey(someMaxPartition, someKind, someNamespace, someName, someTs)
-	keyComparator := NewWatchTableKeyComparator(someKind+"c", someNamespace, someName, zeroData)
-	err := db.View(func(txn badgerwrap.Txn) error {
-		partRes, err1 = wt.GetPreviousKey(txn, curKey, keyComparator)
-		return err1
-	})
-	assert.NotNil(t, err)
-	assert.Equal(t, &WatchTableKey{}, partRes)
-}
